+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - PIM-9771: Fix the image preview when exporting a product as pdf
 - PIM-9822: Fix Error 500 after deleting a filterable asset family attribute
 - PIM-9825: Fix Assets not being properly reindexed after bulk edit
+- PIM-9854: Fix message when deleting a user group used in a project
 - PIM-9853: Make the word "product" translatable 
 
 ## Improvements